Fetch and parse issues for MainGrid
The current code applied to check the origin and target protocol
(http/https) when making a request to avoid problem.

See more information on Stackoverflow as posted by myself here
http://stackoverflow.com/questions/42454963/server-responds-with-400-bad-request-on-accessing-cors-resource
and see #1 for issue of this.

Used static for IssuesManager class as we don't want to waste time in
fetching and parsing all the time.

If client loads an index page, it will take time and parse stuff.
If client refresh the page, it will load from cache, and we only need
time for parsing which is minimal.
If client force clear cache on browser, then everything start off again.
If our resource of issues (json files) are updated, then client will
reload it again even though client didn't clear cache.