Add showing completeness bar graph
Completeness is closed issue / total issues in percentage.
Use react-chartjs-2 to solve the problem of integrating chart.js into
reactjs.

Carefully need to set chartDataSets as one of this.state of reactjs
class in order for it to properly render. This is by meant to avoid
calling forceUpdate().

If you don't add it as one of this.state then when test on iPhone device
(at least this happens) when all latest 5 issues were fetched and
rendered on screen, they will be all white left only bar graph of
completeness.